Add task storage PV parameter to configure mount options
