+++ webserver/app/helpers/const.py
@@ -46,6 +46,7 @@ CRD_DOMAIN = os.getenv("CRD_DOMAIN")
 TASK_REVIEW = os.getenv("TASK_REVIEW")
 TASK_CONTROLLER= os.getenv("TASK_CONTROLLER")
 STORAGE_CLASS = os.getenv("STORAGE_CLASS")
+MOUNT_OPTIONS = os.getenv("MOUNT_OPTIONS")
 GITHUB_DELIVERY = os.getenv("GITHUB_DELIVERY")
 OTHER_DELIVERY = os.getenv("OTHER_DELIVERY")
 ALPINE_IMAGE = os.getenv("ALPINE_IMAGE")
