+++ webserver/app/helpers/task_pod.py
@@ -11,7 +11,7 @@ from kubernetes.client import (
     V1PersistentVolumeClaimSpec, V1VolumeResourceRequirements,
     V1CSIPersistentVolumeSource
 )
-from app.helpers.const import ALPINE_IMAGE, RESULTS_PATH, STORAGE_CLASS, TASK_NAMESPACE
+from app.helpers.const import ALPINE_IMAGE, RESULTS_PATH, STORAGE_CLASS, TASK_NAMESPACE, MOUNT_OPTIONS
 from app.helpers.kubernetes import KubernetesClient
 from app.models.dataset import Dataset
 
@@ -104,7 +104,8 @@ class TaskPod:
         pv_spec = V1PersistentVolumeSpec(
             access_modes=['ReadWriteMany'],
             capacity={"storage": os.getenv("CLAIM_CAPACITY")},
-            storage_class_name=STORAGE_CLASS
+            storage_class_name=STORAGE_CLASS,
+            mount_options=MOUNT_OPTIONS.split(",") if MOUNT_OPTIONS else None
         )
         if os.getenv("AZURE_STORAGE_ENABLED"):
             pv_spec.azure_file=V1AzureFilePersistentVolumeSource(
